+++ CHANGELOG.md
@@ -6,6 +6,17 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 ---
 
+## [3.0.0]
+
+### Updated
+
+- Updated dependencies.
+- Bumped min. version of `@eightshift/ui-components` to 7.0.0.
+
+### Removed
+
+- Removed `__useSmartImageNext` prop from `MediaPicker`.
+
 ## [2.4.0]
 
 ### Added
@@ -295,6 +306,7 @@ Co-authored with @piqusy
 - Initial release.
 
 [Unreleased]: https://github.com/infinum/eightshift-frontend-libs-tailwind/compare/master...HEAD
+[3.0.0]: https://github.com/infinum/eightshift-frontend-libs-tailwind/compare/2.4.0...3.0.0
 [2.4.0]: https://github.com/infinum/eightshift-frontend-libs-tailwind/compare/2.3.0...2.4.0
 [2.3.0]: https://github.com/infinum/eightshift-frontend-libs-tailwind/compare/2.2.0...2.3.0
 [2.2.0]: https://github.com/infinum/eightshift-frontend-libs-tailwind/compare/2.1.1...2.2.0